add lazy loading for image-bearing components
